requirements.txt
pandas>=2.0.0
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0

